Broke Contact.js down to have form in a separate component
